+++ pkg/module/upstream_lister.go
@@ -1,0 +1,13 @@
+package module
+
+import (
+	"context"
+
+	"github.com/gomods/athens/pkg/storage"
+)
+
+// UpstreamLister retrieves a list of available module versions from upstream
+// i.e. VCS, and a Storage backend.
+type UpstreamLister interface {
+	List(ctx context.Context, mod string) (*storage.RevInfo, []string, error)
+}
